backend/requirements.txt
pdf2image==1.10.0
pytesseract==0.3.0
opencv_python
numpy==1.19.4
pandas
pillow==7.0.0
fastapi[all]


